backend/requirements.txt
# AI Application Copilot — Backend Dependencies

# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Settings
pydantic==2.8.2
pydantic-settings==2.4.0

# OpenAI
openai==1.51.0

# Resume parsing
pdfplumber==0.11.4
python-docx==1.1.2

# NLP / fuzzy matching
rapidfuzz==3.9.7
spacy==3.7.6

# Optional: semantic similarity (uncomment if needed)
# sentence-transformers==3.0.1
# scikit-learn==1.5.2

# Dev
python-dotenv==1.0.1
httpx==0.27.2
